Rework SIGCHLD handling, NDMS-227
The current implementation results in an inconsistent state of the structure "children" and
variable "number_of_children" when many children are created and completed(many SIGCHLD signals an once).
This happens when multiple clients are connected,
or the client creates multiple threads
(e.g. PS4 creates ~ 8 threads on startup and then leaves only one).

Changes:
 * Move child's handling to event loop;
 * Add function process_signal_lock, process_signal_unlock.